Move InputType into types index
